+++ src/layers/selection-utils.js
@@ -1,4 +1,5 @@
 import { COORDINATE_SYSTEM } from 'deck.gl';
+import { DataFilterExtension } from '@deck.gl/extensions';
 import SelectionLayer from './SelectionLayer';
 
 /**
@@ -72,26 +73,6 @@ export function getSelectionLayers(
 }
 
 /**
- * Using a color function and a theme name, return a function
- * that mixes a cell color with a theme background color.
- * Reference: https://github.com/bgrins/TinyColor/blob/80f7225029c428c0de0757f7d98ac15f497bee57/tinycolor.js#L701
- * @param {function} colorFunction Returns a color given a cell.
- * @param {number[]} backgroundColor The scatterplot or spatial background color.
- * @returns {function} Returns a color given a cell.
- */
-function mixFunction(colorFunction, backgroundColor) {
-  const p = 0.5;
-  return (cell) => {
-    const rgb = colorFunction(cell);
-    return [
-      ((backgroundColor[0] - rgb[0]) * p) + rgb[0],
-      ((backgroundColor[1] - rgb[1]) * p) + rgb[1],
-      ((backgroundColor[2] - rgb[2]) * p) + rgb[2],
-    ];
-  };
-}
-
-/**
  * Get deck.gl layer props for selection overlays.
  * @param {object} props
  * @returns {object} Object with two properties,
@@ -107,16 +88,19 @@ export function overlayBaseProps(props) {
       id: getSelectedLayerId(id),
       getFillColor: getColor,
       getLineColor: getColor,
-      data: data.filter(isSelected),
+      data,
+      getFilterValue: isSelected,
+      extensions: [new DataFilterExtension({ filterSize: 1 })],
+      filterRange: [0.99, 1.01],
       ...rest,
     },
     base: {
       id: getBaseLayerId(id),
-      getLineColor: getColor ? mixFunction(getColor, backgroundColor) : undefined,
-      getFillColor: getColor ? mixFunction(getColor, backgroundColor) : undefined,
+      getLineColor: getColor,
+      getFillColor: getColor,
       // Alternatively: contrast outlines with solids:
       // getLineColor: getColor,
-      // getFillColor: [255,255,255],
+      // getFillColor: [255, 255, 255],
       data: data.slice(),
       ...rest,
     },